requirements.txt
asgiref==3.11.1
attrs==25.4.0
colorama==0.4.6
coverage==7.13.3
Django==5.2.11
django-environ==0.12.0
django-filter==25.2
django-simple-history==3.11.0
djangorestframework==3.16.1
drf-spectacular==0.29.0
factory_boy==3.3.1
Faker==40.4.0
gunicorn==25.0.2
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==26.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.9.0
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
ruff==0.9.6
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
